requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
aniso8601==9.0.1
apturl==0.5.2
asgiref==3.4.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==1.5.7
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
billiard==4.1.0
blinker==1.4
Brlapi==0.8.3
celery==5.3.1
certifi==2023.5.7
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distlib==0.3.7
distro==1.7.0
distro-info===1.1build1
Django==3.2.12
django-cors-headers==4.1.0
django-crispy-forms==1.13.0
django-crontab==0.7.1
django-debug-toolbar==3.2.2
django-filter==23.5
django-graphql-jwt==0.3.4
django-rest-framework==0.1.0
django-summernote==0.8.11.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
duplicity==0.8.21
executing==1.2.0
fasteners==0.14.1
filelock==3.12.2
Flask==2.2.3
frozenlist==1.3.3
future==0.18.2
graphene==3.2.2
graphene-django==3.1.2
graphql-core==3.2.3
graphql-relay==3.2.0
httplib2==0.20.2
idna==3.4
imageio==2.33.0
imageio-ffmpeg==0.4.9
importlib-metadata==4.6.4
ipdb==0.13.13
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
kombu==5.3.1
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
moviepy==1.0.3
multidict==6.0.4
mysqlclient==2.2.0
netifaces==0.11.0
numpy==1.26.2
oauthlib==3.2.0
olefile==0.46
pandas==2.2.0
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
proglog==0.1.10
promise==2.3
prompt-toolkit==3.0.39
protobuf==3.12.4
ptyprocess==0.7.0
pure-eval==0.2.2
PyAutoGUI==0.9.53
pycairo==1.20.1
pycodestyle==2.8.0
pycups==2.0.1
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.7.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
PyScreeze==0.1.28
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==1.0.0
python3-xlib==0.15
pytweening==1.0.4
pytz==2021.3
pyxdg==0.27
PyYAML==5.4.1
redis==4.6.0
reportlab==3.6.8
requests==2.31.0
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.4.1
SpeechRecognition==3.10.0
sqlparse==0.4.2
stack-data==0.6.2
systemd-python==234
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tqdm==4.66.1
traitlets==5.9.0
twilio==8.3.0
typing_extensions==4.6.3
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.3
usb-creator==0.3.7
vine==5.0.0
virtualenv==20.24.3
wadllib==1.3.6
wcwidth==0.2.6
Werkzeug==2.2.3
wikipedia==1.4.0
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
